feat: Implement currency selection and formatting across the app
- Added CurrencyCubit to manage currency state and preferences.
- Introduced CurrencySelectorTile for selecting the currency.
- Updated transaction-related widgets to display amounts in the selected currency format.
- Modified edit_transaction_sheet to accept double amounts.
- Enhanced home_balance_container to format balance based on selected currency.
- Updated import/export functionality to handle currency formatting.
- Integrated currency formatting in transaction details and tiles.
- Added profile management with image upload capability.

+++ lib/Bloc/transactions/transaction_state.dart
@@ -5,7 +5,7 @@ class TransactionLoading extends TransactionState {}
 class TransactionLoaded extends TransactionState {
   final List<Map<String, Object?>> transactions;
   final Map<String, String> categoryMap;
-  final int cash, card, upi;
+  final double cash, card, upi;
 
   TransactionLoaded(this.transactions, this.categoryMap, {required this.cash, required this.card, required this.upi} );
 }
